pyproject.toml
[project]
name = "agentport"
version = "0.1.0"
description = "Portable AI agents: Letta .af format support, cross-framework migration, CLI & GitHub Actions"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AgentPort Team", email = "team@agentport.dev"}
]
keywords = ["ai", "agent", "letta", "portable", "llm", "agent-framework"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]

[project.scripts]
agentport = "agentport.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentport"]

[tool.hatch.build.targets.sdist]
include = ["src/agentport"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
